+++ controller_interface/include/controller_interface/controller_interface_base.hpp
@@ -50,7 +50,7 @@ enum class return_type : std::uint8_t
  * @enum ALL - Claim all available interfaces.
  * @enum INDIVIDUAL - Claim only the specified individual interfaces.
  * @enum NONE - Claim no interfaces.
- * @enum INDIVIDUAL_FLEXIBLE - Claim the available interfaces from the specified individual
+ * @enum INDIVIDUAL_BEST_EFFORT - Claim the available interfaces from the specified individual
  * interfaces.
  * @enum REGEX - Claim interfaces matching the specified regular expressions.
  */
@@ -59,7 +59,7 @@ enum class interface_configuration_type : std::uint8_t
   ALL = 0,
   INDIVIDUAL = 1,
   NONE = 2,
-  INDIVIDUAL_FLEXIBLE = 3,
+  INDIVIDUAL_BEST_EFFORT = 3,
   REGEX = 10
 };
 
@@ -377,7 +377,7 @@ protected:
    * If interface_configuration_type::ALL is specified, the order is determined by the internal
    * memory of the resource_manager and may not be deterministic. To obtain a consistent order, use
    * \ref get_ordered_interfaces() from \ref helpers.hpp.
-   * If interface_configuration_type::INDIVIDUAL_FLEXIBLE or REGEX is specified, the order might
+   * If interface_configuration_type::INDIVIDUAL_BEST_EFFORT or REGEX is specified, the order might
    * not matched the requested one, as it depends on the available interfaces in the resource
    * manager. Use the \ref get_ordered_interfaces() from \ref helpers.hpp to obtain a consistent
    * order.
@@ -391,7 +391,7 @@ protected:
    * If interface_configuration_type::ALL is specified, the order is determined by the internal
    * memory of the resource_manager and may not be deterministic. To obtain a consistent order, use
    * \ref get_ordered_interfaces() from \ref helpers.hpp.
-   * If interface_configuration_type::INDIVIDUAL_FLEXIBLE or REGEX is specified, the order might
+   * If interface_configuration_type::INDIVIDUAL_BEST_EFFORT or REGEX is specified, the order might
    * not matched the requested one, as it depends on the available interfaces in the resource
    * manager. Use the \ref get_ordered_interfaces() from \ref helpers.hpp to obtain a consistent
    * order.